requirements.txt
bcrypt
docx2txt>=0.8
Flask>=2.1.0
google_auth_oauthlib>=0.4.6
nltk>=3.4.5
numpy>=1.20.2
pandas>=1.4.1
pip>=22.2.2
protobuf
pymongo>=4.2.0
PyMuPDF>=1.20.2
python-dotenv>=0.21.0
requests>=2.22.0
scikit_learn>=1.1.2

spacy>=3.4.1
tensorflow>=2.9.1
Werkzeug>=2.0.3

bcrypt~=4.0.1
scikit-learn~=1.1.3



